tidy: remove unuse or redundant code
